test/e2e: relax pending test check
This test is meant to cause a pending timeout using a non-existent image.  It
works as intended, but it intentionally uses a short timeout period so that it
executes quickly.  It may happen that the timeout happens for other reasons
(this is after all what the actual verification is meant to detect), leading to
test failures such as the one seen here:

https://github.com/openshift/ci-tools/pull/3390#issuecomment-1520443642

Since the reason for the delay is not relevant for this test, only that it be
detected and properly reported, this change relaxes the error message searched
in the logs such that any pending situation satisfies the test.