pkg/kubernetes: demote hack message to debug log
This function is used legitimately in cases where the object is expected to be
deleted as part of regular operation, e.g. watching observer pods:

```
WARN[2023-05-16T13:31:19Z] failed to get object after finishing watch    error=pods "multi-observers-observer" not found
```

https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_ci-tools/3431/pull-ci-openshift-ci-tools-master-e2e/1658459674833326080#1:build-log.txt%3A121

The logging client is not part of any critical process (and has problems of its
own), so this message is not useful for someone looking at the test output.