ci-operator: remove claim release image from input
These changes remove the pull-spec for the release payload image of the cluster
pool used by a test from the inputs used to determine the temporary test
namespace.

---

When a test requests a cluster from a pool (i.e. it has a `cluster_claim`), the
release payload configured for the pool, along with its images, is imported into
the test namespace (see https://github.com/openshift/ci-tools/pull/2074).  This
is done so that tests can use these images in the same way as they would other
types of releases.

---

The original implementation had a problem: since only tests with a
`cluster_claim` are given the Hive `kubeconfig`, other tests in the same
configuration would fail to obtain this information (see
https://github.com/openshift/ci-tools/pull/2092).

https://github.com/openshift/ci-tools/pull/2099 attempted to fix this by only
including the release import step when the client is present.  This has the
dubious effect of executing the test without claiming any cluster if the client
is not provided.  But, more importantly, it introduces another problem: since
the `ImportReleaseStep` declares the pull-spec of the release payload in its
`Inputs` --- as normally we want a separate test namespace if that pull-spec
has changed --- tests with an otherwise equal set of inputs are assigned
different temporary namespaces depending on whether they have a `cluster_claim`
or not.  E.g.:

- https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_hive/2029/pull-ci-openshift-hive-master-e2e/1661129286888722432
- https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_hive/2029/pull-ci-openshift-hive-master-e2e-pool/1661129286913888256
- https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_hive/2029/pull-ci-openshift-hive-master-images/1661129286939054080
- https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_hive/2029/pull-ci-openshift-hive-master-unit/1661129286964219904
- https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_hive/2029/pull-ci-openshift-hive-master-verify/1661129286989385728

All of these jobs have the same set of inputs: they come from the same pull
request, have the same configuration, etc., but the first two are put into a
separate namespace.  While this is not a correctness problem (we could in
theory execute each test in its own namespace), this wastes resources since
none of the work (image builds, etc.) is shared between the two namespaces.

---

This behavior of `ImportReleaseStep` is desired in the case of regular release
payloads: we _want_ to import the new release if the pull-spec has changed, e.g.
if a new version is published.  But in the specific case of cluster pools, it is
not as important.  While some of them are regularly updated to use new payloads,
tests which use pools are not as concerned with using the absolute latest
versions.  Furthermore, it is already possible (in principle, at least) for a
test to import the release payload for a pool and later receive a lease for a
cluster provisioned with a different version.

For these reasons, these changes modify the `ImportReleaseStep` to only
conditionally include its pull-spec in the `Inputs`.  This continues to be done
for all cases, except now when the step is the result of a `cluster_claim`.