Prepare for partitioned Test/Alert/Disruption tables in BigQuery
The TestRUns table is over 200GB dating back more than 2 years. It's costing us
roughly 20$ a day to query it for our test run summary table.

We want to move to partitioned tables which must be done delicately.
Firstly however we need a time column to partition on, which was not
present on TestRows previously.

Each table gains 5 new columns to flatten job run info, into the rows
with the numerical data we track. BigQuery makes storage cheap and
joining expensive.

I also have a query ready that builds a new partitioned table with all
the current data.

The plan is to:

- merge this change and start updating these columns when we upload
- when the uploaders are NOT running:
  - build the new table with a different name
  - rename TestRuns to TestRuns_OLD (repeat for each)
  - rename new table to TestRuns

All columns will then be populated, and new incoming rows will as well.