cincinnati client: query integration & staging instances
OTA maintains staging (identical to production) and integration (running
on engineering candidate OCP clusters) Cincinnati instances. We are
searching for traffic that we could route to especially the integration
one, so that we can find possible problems with engineering candidate
early.

All the instances are serving identical data (up to some minimal skew
coming from when individual instance scrape their source data) so we
should be able to easily use the integration instance first, then
the staging one, and if needed, use the production one as a backstop.