Update historical disruption query
We now use 30 days instead of 21, we believe things are stable enough to
move up and this may help smooth out some of the weekly variance in the
data in this file.

Switched to use the latest daily report instead of a separate copy of
the same query, which was left behind when MasterNodesUpdated was
handled elsewhere.

We now use MasterNodesUpdated != N for the P99 limits, which implies
either Y or null for past releases. We want to use the worst case for
now as we recently did with aggregation. FromRelease empty is
also included as those will always be MasterNodesUpdated=N.