pod-scaler: subscribe, then reload
This corrects a race condition in the current order of operations: the
subscription happens after the loading process begins, which can cause updates
to be missed.  While not a problem in practice, since the loading process is
much longer, it can sometimes cause E2E tests to be delayed:

- https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_ci-tools/3533/pull-ci-openshift-ci-tools-master-e2e/1677005009184624640
- https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/pr-logs/pull/openshift_ci-tools/3533/pull-ci-openshift-ci-tools-master-e2e/1677005009184624640/artifacts/e2e/e2e/build-log.txt

The delay is caused by the readiness probe, which never succeeds since the
reloading interval is currently one hour.  The test still passes because it has
a timeout period of 5m for the probe, after which it will continue executing.

This is all harmless (though the race condition is legitimate), so this change
is mostly about correctness.