ci-operator: allow accessory promotions
There exist repositories which promote images to the OCP payload but
would also like to be able to build utility images that are promoted to
some other location, since inclusion into the core payload ImageStreams
is tightly regulated.

The new `accesssory_promotion` stanza in `ci-operator` configuration
allows users to opt images into other promotion schemes. This field is
opt-in, so no changes are expected for any current configurations.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>