testgrid-config-generator: allow OLM tests
RedHat manages quite a large list of TestGrid configurations that are
*not* derived from the release controller configuration. It's a shame
that every single one of those folks need to manually configure
TestGrid for their use-case and cannot take advantage of this tool and
the automatic PR updates.

This is a small commit that mirrors the OSD additions added last year;
in the process of working through the wrinkles for this new TestGrid
config we will know how to factor the code such that other config could
also be added here.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>