Don't fail aggregation if junit isn't parseable
If any junit file isn't valid XML ten it fails the entire aggregation,
even if it's a small number. We should just log it, but still try to do
the aggregation.

If too many fail, the aggregator will fail at a higher level informing
us it doesn't have enough results.