Remove unnecessary use of BackendDisruption_JobRuns
This data is now carried on every row in BackendDisruption, so there is no need to join in the job runs table we hope to kill off shortly.

One of these queries does jump from 150mb to 1.5gb by using the distinct job runs in BackendDisruption instead, however this is a once per aggregation query and thus only run a few times a day, so not of particular concern.