pkg/steps: don't wait on builds in single-arch contexts
There is a bug in how the build controller chooses the digest of the
image to build from - seemingly logic should be added there to prefer
the multi-arch manifest image when one exists in the IST. The existing
`time.Sleep()` call is a *very* costly mechanism to work around this
bug, and we want to be very, very, very careful about using it, as it
directly translates to developer time. As the workaround is only
pertinent in multi-arch builds, we can at the least omit it in single-
arch contexts.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>