NO-JIRA: Use the branch tip as BashSHA
Follow up an internal Slack conversation [1].
The reason is that the branch tip could be moved forward
while the pull is based on a stale commit. The prow hook
plugin does a similar thing [2].

We do it only for `multi-pr-prow-plugin` because the github
client is available. Leaving a TODO for `prpqr_reconciler`
for the future work if needed.

[1]. https://redhat-internal.slack.com/archives/CBN38N3MW/p1764264237591039

[2]. https://github.com/kubernetes-sigs/prow/blob/db89760fea406dd2813e331c3d52b53b5bcbd140/pkg/plugins/trigger/pull-request.go#L51-L58